// data/github/current/ui/packages/billing-app/types/usage.ts
import type {Products} from '../constants'
import type {UsageGrouping, UsagePeriod} from '../enums'

export interface Filters {
  customer: CustomerSelection
  group: GroupSelection | undefined
  period: PeriodSelection | undefined
  product: Products | undefined
  searchQuery: string
}

export interface CustomerSelection {
  id: string
  displayText: string
}

export interface GroupSelection {
  type: UsageGrouping
  displayText: string
}

export interface PeriodSelection {
  type: UsagePeriod
  displayText: string
}

export interface Usage {
  product: string
  sku: string
  quantity: number
  fullQuantity: number
  billableAmount: number
  unitType: string
}

export interface UsageLineItem {
  entityId: string
  appliedCostPerQuantity: number
  billedAmount: number
  discountAmount?: number
  quantity: number
  fullQuantity: number
  usageAt: string
  totalAmount?: number
  name: string
  dailyLicenseQuantity?: number
  dailyLicenseCost?: number
}

export interface ProductUsageLineItem extends UsageLineItem {
  friendlySkuName: string
  product: string
  sku: string
  unitType: string
}

export interface DiscountUsageLineItem {
  discountAmount: number
  quantity: number
  product: string
  sku: string
  usageAt: string
  unitType: string
}

export interface NetUsageLineItem extends UsageLineItem {
  appliedCostPerQuantity: number
  product: string
  sku: string
  friendlySkuName: string
  quantity: number
  fullQuantity: number
  grossAmount: number
  discountAmount: number
  netAmount: number
  usageAt: string
  unitType: string
  name: string
}

export interface RepoUsageLineItem extends UsageLineItem {
  product: string
  repo: {
    name: string
  }
  org: {
    name: string
    avatarSrc: string
    login: string
  }
  totalAmount: number
  discountAmount: number
}

export interface OtherUsageLineItem {
  billedAmount: number
  netAmount: number
  discountAmount: number
  usageAt: string
}

export type UsageCardVariant = 'org' | 'repo'

export interface UsageReportSelection {
  type: number
  displayText: string
  dateText: string
}

export interface UsageReportRequest {
  period: number
  start?: string
  end?: string
}

export interface UsageChartDataPoint {
  x: number
  y: number
  custom: {
    discountAmount?: number | 'N/A'
    grossAmount: number
    totalAmount?: number | 'N/A'
  }
}

export interface UsageChartData {
  data: UsageChartDataPoint[]
}